requirements.txt
numpy
streamlit
pillow
keras
tensorflow
torch
torchvision
ultralytics